+++ web/tests/stream_events.test.js
@@ -266,6 +266,17 @@ test("update_property", ({override}) => {
         assert.equal(args.sub.stream_id, stream_id);
         assert.equal(args.val, 3);
     }
+
+    // Test stream stream_topic_access_group change event
+    {
+        const stub = make_stub();
+        override(stream_settings_ui, "update_stream_topic_access_group_id", stub.f);
+        stream_events.update_property(stream_id, "stream_topic_access_group", 4);
+        assert.equal(stub.num_calls, 1);
+        const args = stub.get_args("sub", "val");
+        assert.equal(args.sub.stream_id, stream_id);
+        assert.equal(args.val, 4);
+    }
 });
 
 test("marked_unsubscribed (code coverage)", () => {